Allow typical response time calculation to be disabled
Defining MQTTC_NO_TYPICAL_RESPONSE_TIME will disable all code
related to the response time calculation.

Since this value is not used by the library, and is only for
application use, it's useful to have the ability to disable
it when it's not needed. This is particularly the case for
constrained systems with limited memory and no FPU, where the
floating point operations will cause soft-float libraries to
be linked in, increasing binary size.